Downgrade all NoJoinOp related errors that happen during slow boot (#13544)
I've noticed (based on looking at just one session) that NoJoinOp event fired before we finished loading container. This is somewhat wrong, as we pause ops processing during boot, so if it takes too long to load container, we are guaranteed to hit timeout and error. Looking at telemetry, it's obvious that's the main reason we hit these errors in automation.

Downgrading such cases to generic events.