Explicitly fail to compress when the payload is too large and would fail compression anyway (#13495)
## Description

ADO:3049

When the payload is large enough, we would reach a `JSON.stringify`
limit which is expressed as a cryptic `Invalid string length` error with
no additional information. This change makes the limitation more
explicit, until we move to a fully stream-oriented approach to
compressing batches. The limit is currently at around 500 MB per batch
of ops.