Remove side effect of coupling op reentry and compression configs, revert stress tests changes - next (#13620)
OP Reentry in the stress tests is still causing test incidents due to the lifecycle of the container. This requires more refactoring of the stress tests. Until then, these configs should be decoupled and the reentry feature disabled in the stress tests in order to continue having test coverage for compression.