Allow for disabling both compression and chunking via feature gates (#13578)
Even if the features are enabled using the container runtime config
(through code), this change adds the option of disabling the features
(in case of emergency) without code changes.

**Regardless of configs or feature gates, compression/chunking is not
disabled on the receiving end. Any client will be able to process
compressed and or chunked ops. The configs/feature gates are relative to
the producer of the ops, not the consumer.**