Multi-regoin FRS perf test pipeline (#13162)
## Description

We have a reference application called the "scenario runner" that runs a
set of scenarios that utilize the Fluid Framework and Azure Client. Its
purpose is to benchmark the performance and reliability of the Fluid
Framework and Azure Client. Currently the 2 scenarios implemented are
container/doc loading and creating. This specific PR works on top of
that reference app by:

* Mapping Aria logs to the `scenario:runner``namespace
* Extends the test-azure-frs pipeline to run on multiple regions
(westus2, ~~westus3~~, ~~eastus~~, westeurope) to compare performance
across deployments
* Configures the test-azure-frs pipeline to run automatically every day
* Splitting the perf tests into their own Azure pipeline
* Fix where a failed test would show as a failure in the pipeline
* Made all variables configurable via environment variables

Note: At this time the WestUS3 and EastUS are not working and have been
commented out until the issue is resolved

Co-authored-by: Sasha Simic <simic.sasa@gmail.com>