Always report  "InstantiateRuntime" event to telemetry (#13412)
In PR #13305, @vladsud added telemetry to the runtime-instantiation event, but wrapped it in a ReportIfTooLong call.

It would be best to always report the instantiation duration. That way we can always rely on it for telemetry (today, we instead have some of our containers add the telemetry, but not all containers do -- so it would be great to be able to standardize our telemetry to always use the Fluid Framework's reported duration).