Support create fluid container on existing file (#12896)
## Description

What is the problem:

We find the current odsp-driver assumes "creating a new fluid file" is
the only way to create a a new container. However, what we want to
achieve is calling POST snapshot against the existing page file without
the "createNew" option (and without filePath or fileName). The updated
OpStream can detect if the file is a SharePoint Site Page and will
create Fluid storage in the alternative file partition of the file.

What is the ideal solution?
Ideally the OdspDocumentServiceFactoryCore should support creating
container by calling normal post snapshot (without createNew) against
existing file.
## Breaking Changes

> If this introduces a breaking change, please describe the impact and
migration path for existing applications below.
> See
[Breaking-vs-Non-breaking-Changes](https://github.com/microsoft/FluidFramework/wiki/Breaking-vs-Non-breaking-Changes)
for details.
> If there are no breaking changes, delete this section.
Adding new types of OdspError which will only used by new code.

## Reviewer Guidance
Adding a new flow of creating new container in odsp-driver which is by
posing a snapshot to an existing file. The ODSP endpoint will figure out
if the target file support such action and create the container.
The code diverges based on whether the resolved URL contains a file name
(use the old "createNew" flow) or itemId (use the new flow)