GC Tombstone: Separate flags for Load v. Usage (required for recovery) (#13634)
We are working on a mechanism to allow consumers to retrieve and revive
a Tombstoned object (see #13606). In these cases, the object will still
appear tombstoned so we need to allow tombstoned objects to be used (ops
sent/received).

This PR separates two flags where there was previously one. Now it's
`Fluid.GarbageCollection.ThrowOnTombstoneUsage` and
`Fluid.GarbageCollection.ThrowOnTombstoneLoad`

I also made some changes to `sendGCTombstoneEvent` in here.