fix: Handle creation range interleaving (integration) (#13657)
This is mostly a lifted version of #13640 that is ported to the
IdCompressor in fluid-internal/tree.

This PR fixes an issue in the compressor in which the normalizer was
unaware of the creation of unallocated final IDs (via cluster
creation/expansion) and thus incorrectly allocated duplicate eager final
IDs. This scenario occurs when range finalizing is slow (e.g. high
latency network) and results in duplicate IDs being handed to the user
of SharedTree.

This PR will be followed by another that makes the same changes in the
non-experimental compressor.