Refactor zamboniSegments out of MergeTree (#13475)
Changes to decouple the zamboniSegments method from the MergeTree
object, as well as a change to packParent to prevent a node with no
segments after scouring from going through the packing process. I'll be
adding test functionality as well for both zamboniSegments and the
update to packParent. I'm opening this as a draft for now to get
feedback on the refactor and tests before I move forward with the
changes, mainly if I can access MergeTree methods inside of
zamboniSegments (which is now outside of MergeTree) without making all
of them public.

Related to
[AB#1919](https://dev.azure.com/fluidframework/internal/_workitems/edit/1919)
and [PR#13301](https://github.com/microsoft/FluidFramework/pull/13301).


[AB#2817](https://dev.azure.com/fluidframework/internal/_sprints/taskboard/Team%20Curtis/internal/CY23Q1/2023.01/Team%20Curtis%20-%202023.01?workitem=2817)