fix: Handle creation range interleaving (#13640) (#13682)
This PR fixes an issue in the compressor in which the normalizer was
unaware of the creation of unallocated final IDs (via cluster
creation/expansion) and thus incorrectly allocated duplicate eager final
IDs. This scenario occurs when range finalizing is slow (e.g. high
latency network) and results in duplicate IDs being handed to the user
of SharedTree.

Co-authored-by: taylorsw04 <60717813+taylorsw04@users.noreply.github.com>