Attempt to fetch deltas from storage before marking doc as corrupt (#13685)
## Description

There are cases where there is a gap between the incoming op and the op
that scribe is expecting. In these cases, the document gets marked as
corrupt (i.e. scribe ignores all ops from this point and the document
does not attempt to summarize). With this change, we attempt to fetch
these missing ops. Instead of fetching the ops directly from storage, we
use the getdeltas api in alfred since services could have additional
redundancies for op storage and retrieval.

## Reviewer Guidance

Tested this by ignoring a group of ops in the scribe lambda handler and
verified that the situation that would have otherwise resulted in doc
corruption, continued to handle ops without issues