Prevent scribe from creating summaries for incorrect states (#13485)
The following sequence of events causes corruption when scribe is
configured for an external summarizer (`isExternal` is set to true)
1. Summarizer1 sent a summary op with ref seq # 2057
2. Summarizer2 sent a summary op with ref seq # 2056
3. Server processes summary op for ref seq # 2057 and sends a
summaryNack
4. Server processes summary op for ref seq # 2056 and sends a summayAck

When a new client connects and downloads the summary for ref seq # 2056,
the data in the summary has seq#2057 (it should be seq#2056).

This happens because:
1. Scribe does not revert the state for summary failures when an
external summarizer is enabled.
2. Scribe does not have any checks regarding the summary ref sequence
number. Scribe is very good at ensuring it processes ops in order, but
it has no such checks for the summary ref seq num vs scribe's protocol
state.

This PR adds a check for number 2 - it will ensure that scribe will not
make summaries when the summarize op ref sequence number is behind its
current state. Scribe will simply ignore summary ops for this case. We
cannot make it nack for this case due to the potential of kafka message
replays.