r11s: utilize Throttling weight for submitOp and submitSignal (#13648)
## Description

This is a bug fix for op and signal throttling in R11s. The current
implementation does not take op or signal batching into account, so if a
client sends a batch of 10k ops within a `submitOp` event, the
throttling logic will treat it as 1 op. Same for `submitSignal`.

To address this bug, this PR uses the already-implemented throttle
increment weight functionality to increment each submitOp event with
weight equal to total number of messages.

## Reviewer Guidance

### Bad User Experience & Throttling Correctness

I left a note in the comments under `submitOp` about throttling before
checking op sizes. I think this is worth fixing, but might be out of
scope.

### Loop Efficiency

I recognize that looping through the batches to count and then looping
again to order is inefficient, but the alternative would be incrementing
throttle for every message during the order loop, and I don't know if
the client would handle a half-processed batch well if we throttle
mid-processing. Otherwise, we could pre-process the messages, keeping
track of message count there during sanitization, then looping again to
order them, but that's the same efficiency as looping to count, however,
it could solve the NACK inefficiency problem described above.

I also realized my previous op size check fix (#11117) was now
inefficiently looping twice when it only needed to loop once, so I fixed
that here. The filter method their used to be used for other filtering,
but that other filtering was removed in a `Main -> Next` commit
(5564403dce7c620bef8cf94b1552cd9f1600eb6e).