r11s: Add telemetry around throttling internal cache (#13678)
## Description

R11s throttling utilizes an in-memory LRU cache to keep track of local
throttle status in-between computations and syncs with the external
Redis cache. This cache is necessarily limited for memory consumption
purposes, but if the cache is _too_ limited, it can cause problems as
throttling information for different clients can be lost.

AFR has configurations that we think works, but we don't have the
telemetry to back that up. This PR
1. adds telemetry around cache purge for items that have not expired by
age, which is an indication of the cache overflowing.
2. adds telemetry for when a throttler instance begins tracking a
specific key, which will help us confirm that the throttle interval is
being respected consistently.
3. adds an error log where a promise was being marked as `void`
(ignored) previously.