Revert reentry stress tests next - again (#13710)
_The fix in next was overwritten by an integration from main_

Original PR: https://github.com/microsoft/FluidFramework/pull/13620

OP Reentry in the stress tests is still causing test incidents in next due to the lifecycle of the container. This requires more refactoring of the stress tests. Until then, these configs should be decoupled and the reentry feature disabled in the stress tests in order to continue having test coverage for compression.