Fix/property dds/summarization bug
Port #13208 from main to lts branch.

This PR introduce a bug fix for PropertyDDS summarizer, where users
could end-up with inconsistent state in a very spefic scenario described
in more details in the
[test](https://github.com/microsoft/FluidFramework/compare/main...nedalhy:FluidFramework:fix/property_dds/summarization_bug?expand=1#diff-4e672dc85e3021cf5cf96522a3bcf6653dfaaf10639ef2bf0c3de0a12fc8870eR91)
.

This is a fixed version of the PR
https://github.com/microsoft/FluidFramework/pull/13084. I added a fix to
ensure backwards compatibility to summaries that have been generate
before by clients without the fix.

Co-authored-by: Nedal Horany <nedalhy@gmail.com>