Op size benchmark (#13356)
## Description

> Adds a benchmarking test for Operation sizes in utf-8 bytes. The test
benchmarks inserts, deletes and edits of various byte sizes both in
isolation and in different combination. It also will pretty print the
results of all the tests in one table to the console, sorting the
results by average byte size.

The percentiles of utf-8 encoded bytes of JsonableTree used within this
test were what was found to be the maximum size that could be
successfully inserted/deleted/edited using the given node counts (100)
and either individual or singular (bulk) transactions. Note, this was on
a 32 core/64GB RAM Codespaces virtual machine.

Using any larger of a byte size of JsonableTree children causes the
"BatchToLarge" error; this would require either:
Adding artificial wait, for e.x. by using a for-loop to segment our
transactions into batches of less than the given node count OR Making
the size in bytes of the children smaller.


## Example test console output:
![example op size test output
v2](https://user-images.githubusercontent.com/105244057/213243949-e9d934db-cbfb-4660-9675-25a3bd7fe989.png)


## Breaking Changes
> None

## Reviewer Guidance
- Confirm the benchmarking tests cover all the scenarios you would like
to see with Shared Tree.
- Confirm following functions for inserting, deleting and editing work
as you'd expect for this benchmarking test:
> - ` insertNodesWithInvidualTransactions`
> - `insertNodesWithSingleTransaction`
> - `deleteNodesWithInvidualTransactions`
> - `deleteNodesWithSingleTransaction`
> - `editNodesWithInvidualTransactions`
> - `editNodesWithSingleTransaction`
- Note that I limited the byte size for the various tests based on what
would cause a "BatchToLarge" error