Op reentry workaround for stress tests (#13626)
We need to enable op reentry enforcement in the stress tests. 
However, the test infra depends on the task manager to publish ops as
results of other ops. To work around this, we need to schedule the op in
the next turn. But when doing so, the container now has the opportunity
to change the connection state (as the code used to be synchronous and
now it's not). To work around this new problem, we need to make sure
that we don't submit the op unless the container is connected, otherwise
schedule the op for when the container connects.