GC: Added tracking of deleted nodes when sweep is enabled. (#13686)
When sweep is enabled, nodes that are deleted by GC have to be tracked
across summaries. This information will be used by the container runtime
to identify when a deleted object is used and throw appropriate error.
This PR adds tracking of deleted nodes in garbage collector. It does the
following:
1. Makes tombstone feature and sweep mutually exclusive. If sweep will
run, then tombstone feature will be disabled.
2. The ids of nodes that are deleted is written into the GC summary tree
in a "tombstone" blob.
3. Added a function isNodeDeleted that tells whether a given node has
been deleted by GC or not.