[GC-stress] Update totalSendCount to fix tests timing out (#13760)
The tests are timing out again. The issue is that when this happens,
there are no logs to give an idea of what might be happening. Decreasing
the totalSendCount to see if the test passes and we get logs. That might
give idea into what might be causing the timeout. Last time, it was
because of server throttling.