Add a trivial attributor e2e test (#13751)
## Description

This adds a trivial e2e test for the attribution stack, leveraging the
SharedString DDS bookkeeping.

Plans are to expand on this suite to include more reasonable valid
configurations (so that we have good test coverage for integration when
customers go to adopt this API). I'm checking this in early as it also
includes all necessary test infrastructure tweaks to support the
container runtime mixin pattern in test infra, as well as a bug fix for
OpStreamAttributor.