Add SharedCell Attribution (#13560)
## Description

This PR adds functionality and related test suites to SharedCell which supports tracking the
attribution info, i.e. the user who applied the update and the timestamp
when the change happened. The update of attribution info is synchronized
with processing a cell operation, it only happens until an ACK on the
change is received. Therefore, the attribution only works in connected state. 


### PR Checklist

-   [ ] I have updated the documentation accordingly.
-   [X] I have added tests to cover my changes.
-   [X] All new and existing tests passed.
-   [X] My code follows the code style of this project.
- [X] I ran the lint checks which produced no new errors nor warnings
for my changes.
- [X] I have checked to ensure there aren't other open Pull Requests for
the same update/change.

### Does this introduce a breaking change?

-   [ ] Yes
-   [X] No

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>