Fix flaky test in `tinylicious-client` package (#13835)
TinyliciousClient test `can change initialObjects value` is unstable,
because it didn't wait for the op might not have round-tripped thru the
server and the second container to be connected before checking the
values. So the second container might not get the changed state.

The test is introduced in #6405. It is discovered when I ran the test
locally.