Tree: Path prefix (#13766)
## Description

Add an optional prefix to the cursor path methods.

This makes composing cursors easier. For example BasicChunkCursor will
want to delegate to UniformChunkCursor for the leaves of trees, and
needs a way for the paths to compose correctly.

This also helps clear up the semantics of the path getting methods for
non-rooted cases, and provides some functionality that might be
generally useful.

Where possible, cursors implement this functionality directly to avoid
the overhead of wrapping the cursor.
For other cases, a utility is provided to prefix any cursor.

## Breaking Changes

As the tree package is not exported, this does not count as a published
breaking change, however is is breaking at the package API level in two
respects:
1. External implementers of cursor become incorrect (but still build).
Currently there should not be any, so this is fine.
2. Anyone calling getFieldPath or getPath with an argument will
incorrectly pass that as the prefix instead of having it discarded. This
is unlikely to happen with methods but is possible (ex: passing the
result of getPath.bind() to array.map will no longer work)