Additional logging to better understand the GC error from the Stress tests (#13876)
We are investigating a scenario in which, due to the asynchronicity of
our events, causes the summarization to start just before the refresh of
the latest summary gets invoked.
The theory is that we go through the call to refreshSummaryAckLock
(runningSummarizer.ts) before the actual lock is taken by the call to
lockedRefreshSummaryAckAction. We have agreed on the fix, but I'd rather
add a new telemetry event to make sure the investigation is correct,
before proceeding.

This change simply adds a new telemetry event that can be removed later
on when the fix for the actual problem is checked in.