ci(routerlicious): Fix routerlicious CI build (#13889)
Makes the following changes to fix the routerlicious CI (docker) build
now that we're using pnpm in the root of the repo:

- Bump dep on @fluidframework/build-tools to 0.8.0 to pick up latest
build tooling.
- Add missing `assert` dependency in lambdas and memory-orderer. This
worked before because the root dependencies were installed in CI, and
node looks up the tree for deps.
- Excluded `assert` from the lambdas webpack config. Followed the
pattern used in that project for other polyfills.
- Set the working directory to be the build directory when running
genMonorepoJson. This worked before because we just used the root
dependency, but with this change the r1s build will install the
build-tools in the server/routerlicious folder instead, and run it
relative to that folder. It's now correctly isolated from the rest of
the repo, as it should be.