Reconnect message is no longer relevant. (#13887)
## Description

The error message for the reconnection tracker is no longer relevant, as
large ops/batches fail earlier with a different message. Adding another
end-to-end test to showcase this and retiring an obsolete test (we don't
have a max op size anymore). The runtime can still close when it enters
a tight feedback loop with no progress.