[r11s] Adding errorCode property to ConnectDocument metric (#13901)
## Description

In the recent past, our on-call engineers in AFR have had to handle
higher severity (Sev 2) incidents related to `ConnectDocument` failures.
We also had multiple other Sev 3 and Sev 4 incidents of such type. Upon
investigation, it turned out that most of the failures were caused by
errors that are related to the client side, and not server side (e.g.
401, 403, 429).

Even though `ConnectDocument` is not an HTTP metric per se, it can
potentially fail due to HTTP-related errors. For example:

- When processing a connect_document event, an authorization token is
required. If there is no authorization token, we [throw a 403
NetworkError](https://github.com/microsoft/FluidFramework/blob/c1c340fad35a7866923dc60ab5f8f73d8bd97d21/server/routerlicious/packages/lambdas/src/alfred/index.ts#L259).
- If the number of clients connected to the document already reached the
maximum (which is an FRS limit communicated to customers), we [reject
new connect_document requests with
429](https://github.com/microsoft/FluidFramework/blob/c1c340fad35a7866923dc60ab5f8f73d8bd97d21/server/routerlicious/packages/lambdas/src/alfred/index.ts#L347).

Those are client-side errors and the FRS OCE cannot do much about them.
However, the current implementation of the ConnectDocument metric does
not allow for telling those cases apart from real issues. The error
status codes from the examples above are captured in the “exception”
field of the metric, which is a free-form text field that is not
ingested to our metrics platform.

This PR introduces a new optional property for Lumberjack called
`errorCode`, that can be used in the situation described above. That
property can be leveraged in the Lumberjack engine in use, and emitted
as part of the metric data so that metric-based monitors/alerts can have
rules to ignore failures based on certain `errorCode`s. This property is
also flexible and can be used in other situations.