ci(azure): Correct component governance working dirs (#13908)
Our upgrade to pnpm in the root of the repo has caused the component
governance task for the azure release group pipeline to fail because the
task previously assumed it was running from the root.

This PR updates the tasks to run in the azure folder and install
dependencies from there. The pipeline no longer relies on dependencies
from the root.