gitrest: manually reset in-memory volume (#13911)
## Description

In theory, the `memfs` `Volume` object used in the new Gitrest
optimizations from #13249 should be garbage collected by Node.js after
the function ends. However, we see symptoms of a new memory leak in
Gitrest following this change, so I am adding a manual reset of the
created memfs Volume after its usage is complete. We do this in another
service in FRS that uses similar functionality, but I did not think it
would be necessary in Gitrest due to the differences in implementation.