feat: Split mock services into 2: Customer Service and External Data Service (#13542)
This PR updates the current mock service pattern by splitting the
current responsibilities into two separate services:
- The mock external data service (mocking something like Jira, including
its webhook system)
- The mock customer service, which knows about some external data
service and knows how to talk to it.
- This service, for now, emulates the client-local pattern in terms of
how the app consumes it. The Fluid services can register for webhook
notifications, which are echoes of data updates from the external data
service, and can signal the application about the updates as needed.

Mocking other service patterns will occur in subsequent PRs.

This PR also:
- Updates the mock webhook to operate on our new `TaskData` type, and
(currently) passes that data when notifying subscribers. We don't
currently take advantage of this, but we may wish to in the future.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>