ci(azure): Fix component governance paths (#13920)
PR #13908 was an incomplete fix. The generateMonorepoJson tool isn't
very flexible and expects to be run from the root, and always creates
its files in the root of the release group. Thus this PR corrects the
paths and runs from the root as the tool expects.

Fortunately these weird tasks will no longer be needed once release
groups all move to pnpm. That's also why it's not worth fixing the
generateMonorepoJson tool to be more flexible.