refactor(tree): Store repair data in a way that's opaque to Delta (#13921)
The purpose of this change is to refactor the repair data store so that
rather than storing the repair data directly, it stores it in an opaque
handle that can be stored in `Delta`s and used by the forest to retrieve
the data. The format of the repair data may be stored in different ways
between the repair data store and the delta manager vs the forest, so we
don't want to have to convert back and forth between the different
formats.

Changes:
- Uses `chunkTree` to store subtrees (`TreeChunk`s) directly as repair
data.
- `ForestRepairDataStore` returns cursors and they are stored directly
on deltas.
- Repair data is restored using the cursors by the forest rather than
the `ForestRepairDataStore`.
- Adds tests that ensure `ChunkedForest` does not create multiple copies
of the data during the repair data capture and restore process. The
tests are skipped because `BasicChunk` currently does not implement the
necessary property to bypass the extra cloning.