ci(r11s): Use correct working dir in r11s build (#13897)
A follow up to #13889. The generateMonorepoJson tool isn't very flexible
and expects to be run from the root of the repo, and always creates its
files in the root of the release group. Thus this PR corrects the paths
and runs from the root as the tool expects.

Fortunately these weird tasks will no longer be needed once release
groups all move to pnpm. That's also why it's not worth fixing the
generateMonorepoJson tool to be more flexible.

I also updated the list of files that triggers the r11s pipeline in for
PRs. The list now includes `include-generate-notice-steps.yml` so PRs
that update that file will trigger the r11s pipeline as it should.