Update testObjectProvider's allowedErrors to support downgrading log category (#13870)
## Description

See
[AB#2535](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2535)

Any event listed in `allowedErrors` will not cause a test failure if
it's logged. This PR adds a flag to each error to specify whether to
downgrade the event category from error -> generic.

This matters because in addition to test results, we also monitor volume
of error logs generated during e2e test runs, and so some errors we
would not like to trigger that alert. But others we may want to know
about if the volume grows too high.

In other words, by _not_ downgrading, it's equivalent in test telemetry to every error that is explicitly expected - it doessn't fail the test but does appear in the e2e telemetry.