build: Include and format package.json and other projects (#13938)
CI runs have a step that sets the version of packages, which edits the
package.json and changes the whitespace formatting. We have a couple of
different ways we set versions, so the quickest fix seemed to be to add
steps to the "set package version" workflow that:

1. Runs a prettier check on just the package.json files in the release
group BEFORE setting versions.
2. Runs a prettier format on just the package.json files in the release
group AFTER setting versions.

Step 1 is needed because otherwise, badly formatted package.json files
would go undetected, because we force them to be formatted in CI in step
2. The no extraneous files check already ignores package.json files
because they change when the package version changes, so we can't rely
on it as a backstop; we need to check them elsewhere in the pipeline.

An alternative to this workaround is to update our pipelines to use flub
to set versions, and update flub to not touch whitespace. That will take
more time, and this approach seems sufficient.

I also opportunistically updated the scripts and config for the
benchmark project since it was outdated, and I updated the scripts for
the release groups so that they lint and format the files in the root of
the release group. And I updated the .prettierignore file as needed to
exclude stuff.

Finally, there were some examples and PropertyDDS projects that did not
have a config file, so I added one and formatted all the code.