Create prototype SharedMap editor control that just displays primitive data (#13964)
This will be the first step in creating the SharedMap view. To start, we
can just display a readonly table view that lists the keys (always
strings) and the values if they are primitive data only. For
non-primitives, we can just display the kind of data found (if known).


## Sample

<img width="1437" alt="1"
src="https://user-images.githubusercontent.com/114451900/216459611-738b32b2-374b-4804-87c2-119fc5cfbeb6.png">

---------

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>
Co-authored-by: Josmithr <josmithr@microsoft.com>
Co-authored-by: jikim-msft <111468570+jikim-msft@users.noreply.github.com>