build: Add jest and mocha to root dependencies (#13990)
This change adds devDependencies on the mocha and jest testing tools to
the root package.json in the repo. This is to enable easier test running
and debugging using VSCode. However, this makes it possible for projects
to have a 'phantom dependency' on
these tools, so I added a dev note to the readme describing why the root
deps are there and warns about phantom deps.