feat(container-loader): Allow setting the container's logger at load time (#14005)
Cherry-pick of #13779 to the 3.0 release branch.

Developers can now specify a logger that should be used by a container
when it is loaded instead of always using the Loader's logger
(`Loader.services.subLogger`).

Today a container's logger is set based on the logger used in the
Loader. This is not ideal in nested container scenarios, where a
container is loading a data store within another container, which is
itself within another container. Apps/hosts often augment the logger
from the loader when loading a container, and when loading a nested
container, it makes sense to provide an augmented logger to the
container directly rather than using the one from the loader.

This PR enables the scenario by adding a `fluid-base-logger` request
header to the container load request. If set, then the container's
`subLogger` will be set to the header value. If the override isn't set,
then the container will use the loader's logger as it does today.