Add commands to prompts (#13762)
Prompt to release should include commands to release independent
packages/release groups:
[AB#1963](https://dev.azure.com/fluidframework/internal/_sprints/taskboard/Team%20Sasha/internal/CY23Q1/Team%20Sasha%20-%202023.01-A?workitem=1963).

The function `promptToReleaseDeps` did not include commands which would
be displayed when a particular release group such as azure, build-tools,
server or client is not released. It was neither included when
independent packages such as tinylicious, eslint-config-fluid,
protocol-definitions, etc are not released. This PR adds those commands
in the `cmd` section. This makes it easier for the release engineer to
copy the commands directly in the console.