Disable large payload stages (#14023)
## Description

The *:large CI stages create lots of traffic and take a long time to
run. This could cause pipeline delays, as other ODSP/FRS stages are
waiting on the `stress-(frs|odsp)-lock`. This has been running for
almost a week, the results have been captured and configs have been
adjusted to get to a point where they can be integrated with the regular
stress tests run.

Re-enabling chunking and compression in the stress tests will be done by
https://github.com/microsoft/FluidFramework/pull/14000 after testing.

There were multiple problems with the original change and leaving them
on was mostly done to expose them. Most notable being the chunk size was
too small causing the test to send 1000 times more ops (in some cases).

The other issue is that these stages use a lock, so while they run,
other odsp/frs stages cannot. If this stage takes too long, it is going
to slow down the other CI runs.