Use detached attribution key (#13991)
## Description

Uses detached attribution keys in `MergeTree`'s insert-only attribution
policy. This avoids serializing keys with seq: 0.

Note that we still optimize for size for seq-based attribution keys (by
eliding the object type and serializing it as only the `seq`). Though
it's possible people will create large chunks of content in a detached
state, the optimization to combine adjacent and equivalent attribution
keys will still avoid creating large summaries in this case.

The change mostly adds coverage to various testing layers for this.