Switch server release group to pnpm (#13187)
This PR updates the server release group to use pnpm instead of npm.
Summary of changes:

- lerna was bumped to v5 in the server release group. This was needed
because earlier releases only supported yarn workspaces properly. I've
attempted this upgrade without pnpm in the past, but it always failed
(presumably due to npm bugs).
- lerna uses the pnpm workspace config and uses pnpm to execute scripts.
- Several packages had missing deps on node libraries like `assert` and
`events`, so I added those explicitly.
- Some packages had missing non-node deps, like axios and sinon. My
guess is that these worked because of hoisting in the past.
- The Docker image uses debian buster instead of stretch. This was
needed to get a later release of the Python build tools that the server
packages require to build (due to node-rdkafka). The latest releases
available for stretch are too old.