ci(r11s): Remove unneeded pipeline steps post-pnpm (#14043)
Now that we're using pnpm in the server release group, some of the CI
steps are unneeded. There was a special case for r11s before in the
pipeline to generate a lockfile with the full contents of the release
group. That is no longer needed because pnpm's lockfile covers the whole
release group and is automatically detected by the component governance
task.

I tested this by running a manual CI job since the PR runs don't include
component governance.