refactor(container-definitions): Add LoaderHeader enum value to override logger
The LoaderHeader enum in the container-runtime library has a new enum value, `baseLogger`, which is
a header that can be used to override the base logger used by a Fluid Container on load. This is a
breaking change because TypeScript enums are not compatible when new enum values are added.

BREAKING CHANGE: This is a breaking change because TypeScript enums are not compatible when new enum
values are added.
Fixes #AB3350.