build: Add a format:changed script (#14014)
This change adds a `format:changed` script to the root package.json that
uses [pretty-quick](https://github.com/azz/pretty-quick) to run prettier
only on changed files. pretty-quick supports a `--branch` argument that
I set to `main` for the default, but there is also a
`format:changed:next` script to run against next.