Retiring the scribe deltas collection (#14017)
## Description

We currently use two db collections to store the ops. One of the
collections is used to keep track of the logtail (ops since the last
summary) in scribe and the other one is used to keep track of all the
ops. Scribedeltas is used to recover the scribe lambdas in the cases
where a session restarts (either due to a scribe crash or due to a
normal session end and restart). With this PR, we hope to get rid of the
scribe deltas collection and only use the deltas collection. With the
consolidation of deltas and scribe deltas collection, we will
effectively be reducing the RU consumption of the local cosmos db by 80%
of what scribe deltas was consuming (This is approx equal to 20% of all
RU consumption in the ops db)

## Reviewer Guidance

I plan to remove the getDeltasViaAlfred flag and the related code once
this change is well validated

## Other optimizations

1) While summarizations, we now look for the logtails first in the in
the unprocessed ops in memory before fetching it from Alfred or cosmos
db. This should reduce at least one db access per client summary
creation.
2) Not restarting scribe on checkpoint failures
3) Do not fetch ops during the first lambda start as there would be no
pending ops

## Test cases

1) Clicker - start session, click up, close all clients and reopen
2) Clicker - start session, click up, close all clients, wait for
session to end and reopen
3) Clicker - start session, restart scribe while the session is in
progress and make sure the session recovers