Fixed bug - RefreshLatestSummaryAck can end up updating wrong latest reference sequence number (#14052)
## Bug
When refreshLatestSummaryAck is called, if pending state corresponding
to the ack is not found, the latest summary is downloaded. This change
was made fairly recently - [Proposal - refreshLatestSummaryAck's
fetchSnapshotFromStorage always retrieving the latest summary by
NicholasCouri · Pull Request #12313 · microsoft/FluidFramework
(github.com)](https://github.com/microsoft/FluidFramework/pull/12313).

This means that the snapshot that is downloaded may have a reference
sequence number (fetchedSummaryRefSeq) that is greater than the ack's
(summaryRefSeq). However, the latest summary reference sequence number
of summarizer nodes and GC is updated to summaryRefSeq which can lead to
inconsitent state and future summaries may be incorrect.

## Fix
After the latest snapshot is downloaded, use the reference sequence
number of this snapshot to update the state of the summarizer nodes and
GC.

## Telemetry Improvements
Made a couple of improvements to the refresh latest summary telemetry:
- Renamed the telemetry to distinguish between refreshLatestSummaryAck
and refreshLatestSummaryAckFromServer. This will help in analyzing logs.
- Added the reference sequence number and version of the downloaded
summary which was missing when snapshot was downloaded for
refreshLatestSummaryAckFromServer.