GC: Add a "GC Feature Matrix" persisted object for controlling feature exposure over a document's lifetime (#14050)
This introduces a concept of "GC Feature Matrix" to the GC
Metadata that we can use to indicate per file (immutable for its
lifetime) the level of support the file has for different features. This
will allow us to retroactively disable a feature for old files in case
we need to (e.g. if a partner finds a GC-blocking bug during Tombstone phase
they can't/won't fix in a self-repairing way, or any such bug during
Sweep phase).

For now add a specific dimension there `tombstoneGeneration` which is
supplied by the app via an undocumented ContainerRuntimeOption, for the
purpose described above. We will add one for sweep (replacing
`SweepEnabled`) next, and may also migrate GC Enabled determination to
this instead of using GC Feature.